remove duplicated file
